+++ dart/common/detail/SharedLibraryManager.hpp
@@ -37,20 +37,18 @@
 #include <string>
 #include <unordered_map>
 
-#include <boost/filesystem.hpp>
-#include <boost/functional/hash.hpp>
-
 #include "dart/common/Deprecated.hpp"
+#include "dart/common/Filesystem.hpp"
 #include "dart/common/Singleton.hpp"
 
 namespace std {
 
 template <>
-struct hash<boost::filesystem::path>
+struct hash<::dart::common::filesystem::path>
 {
-  size_t operator()(const boost::filesystem::path& p) const
+  size_t operator()(const ::dart::common::filesystem::path& p) const
   {
-    return boost::filesystem::hash_value(p);
+    return ::dart::common::filesystem::hash_value(p);
   }
 };
 
@@ -77,7 +75,7 @@ public:
   /// \deprecated Deprecated in 6.10. Please use load(const std::string&)
   /// instead.
   DART_DEPRECATED(6.10)
-  std::shared_ptr<SharedLibrary> load(const boost::filesystem::path& path);
+  std::shared_ptr<SharedLibrary> load(const common::filesystem::path& path);
 
   /// Loads the shared library with the specified path.
   ///
@@ -94,7 +92,7 @@ protected:
 
 protected:
   /// Map from library path to the library instances.
-  std::unordered_map<boost::filesystem::path, std::weak_ptr<SharedLibrary>>
+  std::unordered_map<common::filesystem::path, std::weak_ptr<SharedLibrary>>
       mLibraries;
   // TODO(JS): Remove this in DART 7.