+++ dart/utils/mjcf/detail/Body.cpp
@@ -46,7 +46,7 @@ namespace detail {
 //==============================================================================
 Errors Body::read(
     tinyxml2::XMLElement* element,
-    const common::optional<Size>& size,
+    const std::optional<Size>& size,
     const Defaults& defaults,
     const Default* currentDefault)
 {