+++ dart/utils/mjcf/detail/Geom.cpp
@@ -92,7 +92,7 @@ Errors Geom::read(
 
 //==============================================================================
 static bool canUseFromTo(
-    GeomType type, const common::optional<Eigen::Vector6d>& fromto)
+    GeomType type, const std::optional<Eigen::Vector6d>& fromto)
 {
   if (!fromto)
     return false;