+++ dart/utils/mjcf/detail/WeldAttributes.hpp
@@ -46,12 +46,12 @@ namespace detail {
 
 struct WeldAttributes final
 {
-  common::optional<std::string> mName;
+  std::optional<std::string> mName;
   bool mActive{true};
   Eigen::Vector2d mSolRef;
   Eigen::Matrix<double, 5, 1> mSolImp;
   std::string mBody1;
-  common::optional<std::string> mBody2;
+  std::optional<std::string> mBody2;
   Eigen::Matrix<double, 7, 1> mRelPose;
 
   WeldAttributes();