+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   * Added spdlog support as underlying logging framework: [#1633](https://github.com/dartsim/dart/pull/1633)
   * Added custom memory allocators: [#1636](https://github.com/dartsim/dart/pull/1636), [#1637](https://github.com/dartsim/dart/pull/1637), [#1639](https://github.com/dartsim/dart/pull/1639), [#1645](https://github.com/dartsim/dart/pull/1645), [#1646](https://github.com/dartsim/dart/pull/1646)
   * Added Stopwatch class to replace Timer: [#1638](https://github.com/dartsim/dart/pull/1638)
+  * Removed use of boos::filesystem and boost::optional: [#1648](https://github.com/dartsim/dart/pull/1648)
 
 * Dynamics